+++ SplitWiser/src/main/java/com/splitwiser/SplitWiser/group/GroupRepository.java
@@ -3,8 +3,7 @@ package com.splitwiser.SplitWiser.group;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-
 @Repository
-public interface GroupRepository extends JpaRepository<Group, Long> {
+public interface GroupRepository extends JpaRepository<Group, Integer> {
     // define queries here
 }
